+++ cmd/substreams/protogen.go
@@ -33,7 +33,10 @@ func init() {
 		(e.g. a local file ending with .yaml), the output path will be made relative to it
 	`))
 	protogenCmd.Flags().StringArrayP("exclude-paths", "x", []string{}, "Exclude specific files or directories, for example \"proto/a/a.proto\" or \"proto/a\"")
-	protogenCmd.Flags().Bool("generate-mod-rs", false, "Generate the protobuf mod.rs file")
+	protogenCmd.Flags().Bool("generate-mod-rs", true, cli.FlagDescription(`
+		Generate the protobuf 'mod.rs' file alongside the rust bindings. Include '--generate-mod-rs=false' If you wish to disable this generation.
+		If there is a present 'buf.gen.yaml', consult https://github.com/neoeinstein/protoc-gen-prost/blob/main/protoc-gen-prost-crate/README.md to add 'mod.rs' generation functionality.
+	`))
 }
 
 func runProtogen(cmd *cobra.Command, args []string) error {